Numpy 1.20 compatibility fixes

+++ dask_ml/model_selection/utils_test.py
@@ -184,7 +184,7 @@ class CheckingClassifier(BaseEstimator, ClassifierMixin):
     def predict(self, T):
         if self.check_X is not None:
             assert self.check_X(T)
-        return self.classes_[np.zeros(_num_samples(T), dtype=np.int)]
+        return self.classes_[np.zeros(_num_samples(T), dtype=np.int32)]
 
     def score(self, X=None, Y=None):
         if self.foo_param > 1:
